requirements.txt
black==23.1.0
certifi==2022.12.7       
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Login==0.6.2       
Flask-SQLAlchemy==3.0.3  
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2      
jgo==1.0.5
Jinja2==3.1.2
JPype1==1.4.1
MarkupSafe==2.1.2        
mypy-extensions==1.0.0   
numpy==1.24.2
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0      
psutil==5.9.4
python-dotenv==0.21.1    
rdkit-pypi==2022.9.4     
requests==2.28.2
SQLAlchemy==2.0.2        
tomli==2.0.1
typing_extensions==4.4.0 
urllib3==1.26.14
Werkzeug==2.2.2